+++ src/trail/index.ts
@@ -2,13 +2,17 @@ export {
   EXAMPLE_AGENTIC_RHS,
   BANK_ORDER,
   FIL_MCP_DISPLAY_ONLY,
+  DUST_FLOORS,
   LOW_CAP_SLOW,
+  DIVIDEND_15M,
+  SURF_LEARN,
+  DEFAULT_GATE_MODE,
   TRAIL_HORIZON_DEFAULT,
   FOLLOW_PATH_SOURCES,
   FOLLOW_PATH_STATUSES,
   TRICK_IDS,
 } from "./constants.js";
-export type { BankSymbol, TrickId } from "./constants.js";
+export type { BankSymbol, GateMode, GateProfile, TrickId } from "./constants.js";
 export type {
   FollowPath,
   FollowPathSource,
@@ -26,6 +30,8 @@ export type {
   TroughWindow,
   WatchCandidate,
   WatchResult,
+  WhatIfPath,
+  SurfLearnResult,
 } from "./types.js";
 export {
   baseSymbol,
@@ -35,9 +41,11 @@ export {
   expectancyHaltActive,
   findQuote,
   findSleeve,
+  gateProfile,
   isAgenticAccount,
   isBankSymbol,
   isFilDisplayOnly,
+  liveMicroBuyingPowerOk,
   maxTakeWithoutFlatten,
   quoteSpread,
   refuseNewWorkingEntry,
@@ -50,6 +58,7 @@ export {
 } from "./gates.js";
 export { SuccessLedger, LedgerError } from "./ledger.js";
 export { watch15m } from "./watcher.js";
+export { runSurfLearn, whatIfPnlUsd } from "./surf-learn.js";
 export { FixtureBroker } from "./broker.js";
 export type { TrailBrokerPort } from "./broker.js";
 export { assertSnapshot, SnapshotError } from "./snapshot.js";